+++ src/main/java/rest/request/Coordinates.java
@@ -7,33 +7,33 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class Coordinates {
 
     @JsonProperty("latitude")
-    private Float latitude;
+    private Double latitude;
 
     @JsonProperty("longitude")
-    private Float longitude;
+    private Double longitude;
 
     public Coordinates(){
 
     }
 
-    public Coordinates(Float latitude, Float longitude){
+    public Coordinates(Double latitude, Double longitude){
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
